codemonitors: Make store aware of webhook and slack_webhook columns (#27479)
In #27474, I updated the database schema to add webhook and slack_webhook
columns and tables to code monitors. In this PR, I update the store to be aware of those
types, and add ListActionJobs and CountActionJobs methods that handle listing
and counting for more than just emails.

+++ enterprise/cmd/frontend/internal/codemonitors/resolvers/resolvers.go
@@ -836,7 +836,7 @@ func (m *monitorEmail) Events(ctx context.Context, args *graphqlbackend.ListEven
 	for i, aj := range ajs {
 		events[i] = &monitorActionEvent{Resolver: m.Resolver, ActionJob: aj}
 	}
-	return &monitorActionEventConnection{events: events, totalCount: totalCount}, nil
+	return &monitorActionEventConnection{events: events, totalCount: int32(totalCount)}, nil
 }
 
 //
